Remove separate ```ids``` array, and entered duplicate user checking into the main queue.
Changed main ```queues``` Map from Map<string, Array> to Map<string, Match>.
